+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/ChatMember/LeftChatMemberImpl.kt
@@ -1,12 +1,4 @@
 package dev.inmo.tgbotapi.types.ChatMember
 
-import dev.inmo.tgbotapi.types.*
-import dev.inmo.tgbotapi.types.ChatMember.abstracts.LeftChatMember
-import kotlinx.serialization.*
-
-@Serializable
-data class LeftChatMemberImpl(@SerialName(userField) override val user: User) : LeftChatMember {
-    @SerialName(statusField)
-    @Required
-    private val type: String = "left"
-}
+@Deprecated("LeftChatMemberImpl", ReplaceWith("LeftChatMemberImpl", "dev.inmo.tgbotapi.types.chat.member.LeftChatMemberImpl"))
+typealias LeftChatMemberImpl = dev.inmo.tgbotapi.types.chat.member.LeftChatMemberImpl