+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/ChatMember/MemberChatMemberImpl.kt
@@ -1,12 +1,4 @@
 package dev.inmo.tgbotapi.types.ChatMember
 
-import dev.inmo.tgbotapi.types.*
-import dev.inmo.tgbotapi.types.ChatMember.abstracts.MemberChatMember
-import kotlinx.serialization.*
-
-@Serializable
-data class MemberChatMemberImpl(@SerialName(userField) override val user: User) : MemberChatMember {
-    @SerialName(statusField)
-    @Required
-    private val type: String = "member"
-}
+@Deprecated("MemberChatMemberImpl", ReplaceWith("MemberChatMemberImpl", "dev.inmo.tgbotapi.types.chat.member.MemberChatMemberImpl"))
+typealias MemberChatMemberImpl = dev.inmo.tgbotapi.types.chat.member.MemberChatMemberImpl