+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/ChatMember/abstracts/AdministratorChatMember.kt
@@ -1,25 +1,7 @@
 package dev.inmo.tgbotapi.types.ChatMember.abstracts
 
-import dev.inmo.tgbotapi.utils.RiskFeature
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
+@Deprecated("AdministratorChatMember", ReplaceWith("AdministratorChatMember", "dev.inmo.tgbotapi.types.chat.member.AdministratorChatMember"))
+typealias AdministratorChatMember = dev.inmo.tgbotapi.types.chat.member.AdministratorChatMember
 
-@Serializable(AdministratorChatMemberSerializer::class)
-interface AdministratorChatMember : SpecialRightsChatMember, ChatAdministratorRights {
-    val canBeEdited: Boolean
-    val customTitle: String?
-
-    val canManageVoiceChats: Boolean
-        get() = canManageVideoChats
-}
-
-@RiskFeature
-object AdministratorChatMemberSerializer : KSerializer<AdministratorChatMember> {
-    override val descriptor: SerialDescriptor = ChatMemberSerializer.descriptor
-
-    override fun deserialize(decoder: Decoder): AdministratorChatMember = ChatMemberSerializer.deserialize(decoder) as AdministratorChatMember
-    override fun serialize(encoder: Encoder, value: AdministratorChatMember) = ChatMemberSerializer.serialize(encoder, value)
-}
+@Deprecated("AdministratorChatMemberSerializer", ReplaceWith("AdministratorChatMemberSerializer", "dev.inmo.tgbotapi.types.chat.member.AdministratorChatMemberSerializer"))
+typealias AdministratorChatMemberSerializer = dev.inmo.tgbotapi.types.chat.member.AdministratorChatMemberSerializer