+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/ChatMember/abstracts/MemberChatMember.kt
@@ -1,3 +1,4 @@
 package dev.inmo.tgbotapi.types.ChatMember.abstracts
 
-interface MemberChatMember : ChatMember
+@Deprecated("MemberChatMember", ReplaceWith("MemberChatMember", "dev.inmo.tgbotapi.types.chat.member.MemberChatMember"))
+typealias MemberChatMember = dev.inmo.tgbotapi.types.chat.member.MemberChatMember