+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/BannedChatMember.kt
@@ -1,0 +1,7 @@
+package dev.inmo.tgbotapi.types.chat.member
+
+import dev.inmo.tgbotapi.CommonAbstracts.types.UntilDate
+import kotlinx.serialization.Serializable
+
+@Serializable(ChatMemberSerializer::class)
+sealed interface BannedChatMember : ChatMember, UntilDate