+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/MemberChatMember.kt
@@ -1,0 +1,6 @@
+package dev.inmo.tgbotapi.types.chat.member
+
+import kotlinx.serialization.Serializable
+
+@Serializable(ChatMemberSerializer::class)
+sealed interface MemberChatMember : ChatMember